Savitha/src/com/test/WordReverseTest.java
package com.test;

import org.testng.annotations.Test;

import com.strings.WordReverse;

import static org.testng.AssertJUnit.assertEquals;


public class WordReverseTest {
  @Test(groups = "Positive")
  public void Wordreversetest() {
	  
	  String str = "sihT si eht gnirts hcihw si deserever";
	  String Expected = " This  is  the  string  which  is reveresed";
	  String Actual = WordReverse.reverseWordByWord(str);
	  assertEquals(Actual, Expected);
	}

  @Test(groups = "Negative")
  public void Wordreversetestneg() {
	  
	  String str = "sihT si eht gnirts hcihw si deserever";
	  String Expected = " This is the string which is reveresed";
	  String Actual = WordReverse.reverseWordByWord(str);
	  assertEquals(Actual, Expected);
	}
}
